rename securitypolicy to uapolicy
