+++ uapolicy/policyAes256Sha256RsaPss.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a MIT-style license that can be
 // found in the LICENSE file.
 
-package securitypolicy
+package uapolicy
 
 import (
 	"crypto"
@@ -10,8 +10,8 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/gopcua/opcua/securitypolicy/cipher"
-	"github.com/gopcua/opcua/securitypolicy/sign"
+	"github.com/gopcua/opcua/uapolicy/cipher"
+	"github.com/gopcua/opcua/uapolicy/sign"
 
 	// Force compilation of required hashing algorithms, although we don't directly use the packages
 	_ "crypto/sha1"
