+++ uapolicy/policyPubSubAes128CTR.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a MIT-style license that can be
 // found in the LICENSE file.
 
-package securitypolicy
+package uapolicy
 
 /*
 "SecurityPolicy [A] - PubSub-Aes128-CTR" Profile
